+++ src/apps/toolbar/modules/shared/store/infra.ts
@@ -73,6 +73,7 @@ export async function registerStoreEvents() {
   });
 
   await listenGlobal<MediaChannelTransportData[]>('media-sessions', (event) => {
+    console.log(event.payload);
     store.dispatch(RootActions.setMediaSessions(event.payload));
   });