+++ src/background/modules/media/domain.rs
@@ -13,11 +13,20 @@ use windows::{
 };
 
 #[derive(Debug, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct MediaPlayerOwner {
+    pub name: String,
+    pub icon_path: Option<PathBuf>,
+}
+
+#[derive(Debug, Serialize)]
+#[serde(rename_all = "camelCase")]
 pub struct MediaPlayer {
     pub id: String,
     pub title: String,
     pub author: String,
     pub thumbnail: Option<PathBuf>,
+    pub owner: Option<MediaPlayerOwner>,
     pub playing: bool,
     pub default: bool,
 }